+++ src/select.js
@@ -1,49 +1,39 @@
-export function selectHistoryState(state) {
-  return state.history
-}
-export function getKey(state, keyId) {
-  const key = state.key[keyId]
-  if (!key) throw new Error(`Missing history for key ${key}.`)
-  return key
-}
-export function selectKey(keyType) {
-  return (state, emptyKey = null) => {
-    const keyId = state[keyType]
-    if (keyId === null) return emptyKey
-    return getKey(state, keyId)
-  }
-}
-export const selectActiveKey = selectKey('activeKey')
-export const selectFirstKey = selectKey('firstKey')
-export const selectLastKey = selectKey('lastKey')
-export function getIndex(selector) {
-  return state => selector(state, { index: 0 }).index
-}
-export const getFirstIndex = getIndex(selectFirstKey)
-export const getKeyIndex = getIndex(selectActiveKey)
-export const getLastIndex = getIndex(selectLastKey)
-export function selectActiveKeyDefault(state) {
-  return selectActiveKey(selectHistoryState(state))
-}
+import { partial, partialRight } from 'lodash'
+import { add, find, flow, get, getOr, orderBy, reverse } from 'lodash/fp'
+import { getSelect } from 'cape-select'
+import { createSelector } from 'reselect'
+
+// NOTE: These all act on history state directly. Redux slice needs to be selected first.
+
+export const getHistoryItems = get('key')
+export const getActiveId = get('activeKey')
+export const getFirstId = get('firstKey')
+export const getLastId = get('lastKey')
+
+export const getHistoryItem = partial(getSelect, getHistoryItems)
+
+export const selectActive = getHistoryItem(getActiveId)
+export const selectActiveKey = selectActive
+export const selectFirst = getHistoryItem(getFirstId)
+export const selectFirstKey = selectFirst
+export const selectLast = getHistoryItem(getLastId)
+export const selectLastKey = selectLast
+
+export const getIndex = partialRight(flow, getOr(0, 'index'))
+
+export const selectActiveIndex = getIndex(selectActive)
+export const getKeyIndex = selectActiveIndex
+export const getFirstIndex = getIndex(selectFirst)
+export const getLastIndex = getIndex(selectLast)
+
+export const selectPrevIndex = flow(selectActiveIndex, add(-1))
+export const selectNextIndex = flow(selectActiveIndex, add(1))
+export const getLength = selectNextIndex
+
 export function historyMatch(reduxHistory, windowHistory) {
   if (!windowHistory || !windowHistory.id) throw new Error('Missing window.history.state')
   const activeKey = selectActiveKey(reduxHistory)
   return activeKey.id === windowHistory.id
-}
-export function keyMatch(reduxHistory, windowHistory) {
-  return reduxHistory.activeKey === windowHistory.id
-}
-export function getLength(reduxHistory) {
-  if (!reduxHistory.lastKey) return 0
-  return getLastIndex(reduxHistory) + 1
-}
-export function activeLastMatch({ activeKey, lastKey }) {
-  return activeKey === lastKey
-}
-export function isNewHistory(reduxHistory) {
-  // const length = getLength(reduxHistory)
-  // console.log(browserState.length, length)
-  return activeLastMatch(reduxHistory)// && browserState.length === (length - 1)
 }
 export function browserHistory(reduxHistory) {
   return {
@@ -54,6 +44,12 @@ export function browserHistory(reduxHistory) {
 export function lengthMatch(reduxHistory, windowHistory) {
   return getLength(reduxHistory) === windowHistory.length
 }
-export function browserHasHistory(windowHistory) {
-  return windowHistory.state && windowHistory.state.id
+export const sortedHistoryItems = createSelector(
+  getHistoryItems,
+  orderBy('index', 'asc')
+)
+export const sortedHistoryItemsDesc = flow(sortedHistoryItems, reverse)
+export const selectPrevious = getSelect(sortedHistoryItems, selectPrevIndex)
+export function findDifferentPage(isDifferentFunc) {
+  return flow(sortedHistoryItemsDesc, find(isDifferentFunc))
 }